# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arc LMS API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Arc LMS Support",
      "email": "support@arc-lms.com",
      "url": "https://arc-lms.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://arc-lms.onrender.com/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/classes": {
      "get": {
        "tags": ["Classes"],
        "summary": "List classes",
        "operationId": "listClasses",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/Limit"}
        ],
        "responses": {
          "200": {
            "description": "Classes retrieved",
            "content": {"application/json": {"schema": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/Class"}}, "pagination": {"$ref": "#/components/schemas/PaginatedResult"}}}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "post": {
        "tags": ["Classes"],
        "summary": "Create class",
        "operationId": "createClass",
        "security": [{"BearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateClassRequest"}}}},
        "responses": {
          "201": {"description": "Class created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Class"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/classes/{id}": {
      "get": {
        "tags": ["Classes"],
        "summary": "Get class by ID",
        "operationId": "getClass",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Class retrieved", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Class"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "put": {
        "tags": ["Classes"],
        "summary": "Update class",
        "operationId": "updateClass",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateClassRequest"}}}},
        "responses": {
          "200": {"description": "Class updated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Class"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "delete": {
        "tags": ["Classes"],
        "summary": "Delete class",
        "operationId": "deleteClass",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "204": {"description": "Class deleted"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /public/auth/login"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "description": "Unique key to ensure idempotent requests. Prevents duplicate operations.",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-indexed)",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "description": "Number of items per page",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or token invalid/expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required",
                "details": {}
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "Insufficient permissions to perform this action",
                "details": {}
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "Resource not found",
                "details": {}
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Class": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "grade_level": {"type": "string"},
          "section": {"type": "string"},
          "capacity": {"type": "integer"},
          "form_tutor_id": {"type": "string", "format": "uuid"},
          "status": {"type": "string", "enum": ["ACTIVE", "ARCHIVED"]},
          "student_count": {"type": "integer"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "CreateClassRequest": {
        "type": "object",
        "required": ["name", "grade_level"],
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 100},
          "grade_level": {"type": "string", "minLength": 1, "maxLength": 50},
          "section": {"type": "string", "maxLength": 20},
          "capacity": {"type": "integer", "minimum": 1, "maximum": 100},
          "form_tutor_id": {"type": "string", "format": "uuid"}
        }
      },
      "UpdateClassRequest": {
        "type": "object",
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 100},
          "grade_level": {"type": "string", "minLength": 1, "maxLength": 50},
          "section": {"type": "string", "maxLength": 20},
          "capacity": {"type": "integer", "minimum": 1, "maximum": 100},
          "form_tutor_id": {"type": "string", "format": "uuid"}
        }
      },
...
      "PaginatedResult": {
        "type": "object",
        "properties": {
          "total": {"type": "integer"},
          "page": {"type": "integer"},
          "limit": {"type": "integer"},
          "total_pages": {"type": "integer"},
          "has_next": {"type": "boolean"},
          "has_prev": {"type": "boolean"},
          "next_cursor": {"type": "string"},
          "prev_cursor": {"type": "string"}
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error context",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
    }
  }
}